pyproject.toml
[tool.poetry]
name = "rainbow-pony-api"
version = "0.1.0"
description = ""
authors = ["Renee Lu <reneelu419@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
asyncpg = "^0.30.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
